Improve script.

+++ sloccount.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
 
-TARGET=target/sloccount.sc
+BUILD_DIR="target"
+TARGET="${BUILD_DIR}/sloccount.sc"
 
-rm -rf "${TARGET}"
+if [ -e "${TARGET}" ] ; then
+    rm -rf "${TARGET}"
+fi
+
+if [ ! -d "${BUILD_DIR}" ] ; then
+    mkdir -p "${BUILD_DIR}"
+fi
 
 sloccount --duplicates --wide --details --append \
     commons-application/src/main/java \
@@ -11,10 +18,12 @@ sloccount --duplicates --wide --details --append \
     commons-experimental/src/main/java \
     commons-guava/src/main/java \
     commons-jcommander/src/main/java \
+    commons-parse/src/main/java \
     commons-shell/src/main/java \
     commons-string/src/main/java \
     commons-swing/src/main/java \
     commons-system/src/main/java \
     commons-testing/src/main/java \
     commons-time/src/main/java \
-    commons-validate/ > "${TARGET}"
+    commons-uri/src/main/java \
+    commons-validate/src/main/java > "${TARGET}"
